requirements.txt
attrs==21.2.0
aws-cdk.assets
aws-cdk.aws-acmpca
aws-cdk.aws-apigateway
aws-cdk.aws-applicationautoscaling
aws-cdk.aws-autoscaling
aws-cdk.aws-autoscaling-common
aws-cdk.aws-autoscaling-hooktargets
aws-cdk.aws-certificatemanager
aws-cdk.aws-cloudformation
aws-cdk.aws-cloudfront
aws-cdk.aws-cloudfront-origins
aws-cdk.aws-cloudwatch
aws-cdk.aws-codebuild
aws-cdk.aws-codecommit
aws-cdk.aws-codedeploy
aws-cdk.aws-codeguruprofiler
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-codestarnotifications
aws-cdk.aws-cognito
aws-cdk.aws-ec2
aws-cdk.aws-ecr
aws-cdk.aws-ecr-assets
aws-cdk.aws-ecs
aws-cdk.aws-efs
aws-cdk.aws-eks
aws-cdk.aws-elasticloadbalancing
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-globalaccelerator
aws-cdk.aws-iam
aws-cdk.aws-kinesis
aws-cdk.aws-kinesisfirehose
aws-cdk.aws-kms
aws-cdk.aws-lambda
aws-cdk.aws-logs
aws-cdk.aws-rds
aws-cdk.aws-route53
aws-cdk.aws-route53-targets
aws-cdk.aws-s3
aws-cdk.aws-s3-assets
aws-cdk.aws-sam
aws-cdk.aws-secretsmanager
aws-cdk.aws-servicediscovery
aws-cdk.aws-signer
aws-cdk.aws-sns
aws-cdk.aws-sns-subscriptions
aws-cdk.aws-sqs
aws-cdk.aws-ssm
aws-cdk.aws-stepfunctions
aws-cdk.cloud-assembly-schema
aws-cdk.core
aws-cdk.custom-resources
aws-cdk.cx-api
aws-cdk.lambda-layer-awscli
aws-cdk.lambda-layer-kubectl
aws-cdk.lambda-layer-node-proxy-agent
awscli==1.22.24
botocore==1.23.24
cattrs==1.8.0
colorama==0.4.3
constructs==3.3.166
decorator==5.1.0
docutils==0.15.2
jmespath==0.10.0
joblib==1.1.0
jsii==1.55.1
numpy==1.22.0
publication==0.0.3
pyasn1==0.4.8
PyMySQL==1.0.2
python-dateutil==2.8.2
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==1.0.2
scipy==1.7.3
self==2020.12.3
six==1.16.0
threadpoolctl==3.0.0
typing_extensions==4.0.1
urllib3==1.26.7
